Use local vgg file if present, otherwise download from S3

+++ app/make_image.py
@@ -22,11 +22,15 @@ def make_image(num_iterations, style_image_file, content_image_file=None):
 
     # clear out previous temporary images
     for im_file in os.listdir(TMP_DIR):
-        os.remove(os.path.join(TMP_DIR, im_file))
+        if im_file != '.gitignore':
+            os.remove(os.path.join(TMP_DIR, im_file))
 
-    VGG_MODEL_ADDRESS = 'https://s3.amazonaws.com/style-transfer-playground-assets/imagenet-vgg-verydeep-19.mat'
-    VGG_MODEL_FILE = './tmp/imagenet-vgg-verydeep-19.mat'
-    urllib.request.urlretrieve(VGG_MODEL_ADDRESS, VGG_MODEL_FILE)
+    if os.path.exists('imagenet-vgg-verydeep-19.mat'):
+        VGG_MODEL_FILE = 'imagenet-vgg-verydeep-19.mat'
+    else:
+        VGG_MODEL_ADDRESS = 'https://s3.amazonaws.com/style-transfer-playground-assets/imagenet-vgg-verydeep-19.mat'
+        VGG_MODEL_FILE = './imagenet-vgg-verydeep-19.mat'
+        urllib.request.urlretrieve(VGG_MODEL_ADDRESS, VGG_MODEL_FILE)
 
     vgg = scipy.io.loadmat(VGG_MODEL_FILE)
     vgg_layers = vgg['layers']
